refactor: move debug-jitter to reth-engine-primitives
Move the jitter utility from reth-trie-parallel to reth-engine-primitives
so it can be reused in other components like payload validation and
prewarm workers.

- Create jitter module in reth-engine-primitives
- Update reth-trie-parallel to re-export from engine-primitives
- Rename env var from RETH_PROOF_WORKER_JITTER_MS to RETH_DEBUG_JITTER_MS
- Simplify API to take a single context string
- Propagate debug-jitter feature through the crate hierarchy

Amp-Thread-ID: https://ampcode.com/threads/T-019c3416-f509-7642-94c8-0e2455861c70
