+++ crates/engine/tree/Cargo.toml
@@ -116,7 +116,7 @@ name = "state_root_task"
 harness = false
 
 [features]
-debug-jitter = ["reth-trie-parallel/debug-jitter"]
+debug-jitter = ["reth-engine-primitives/debug-jitter", "reth-trie-parallel/debug-jitter"]
 test-utils = [
     "reth-chain-state/test-utils",
     "reth-chainspec/test-utils",
