+++ crates/trie/parallel/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 [dependencies]
 # reth
 reth-primitives-traits = { workspace = true, features = ["dashmap"] }
+reth-engine-primitives = { workspace = true, optional = true }
 reth-execution-errors.workspace = true
 reth-provider.workspace = true
 reth-storage-errors.workspace = true
@@ -40,8 +41,7 @@ crossbeam-channel.workspace = true
 reth-metrics = { workspace = true, optional = true }
 metrics = { workspace = true, optional = true }
 
-# `debug-jitter` feature - random delays for testing timing-related bugs
-rand = { workspace = true, optional = true }
+
 
 [dev-dependencies]
 # reth
@@ -60,7 +60,7 @@ tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
 [features]
 default = ["metrics"]
 metrics = ["reth-metrics", "dep:metrics", "reth-trie/metrics", "reth-trie-sparse/metrics"]
-debug-jitter = ["dep:rand"]
+debug-jitter = ["reth-engine-primitives/debug-jitter"]
 test-utils = [
     "reth-primitives-traits/test-utils",
     "reth-provider/test-utils",
