+++ crates/trie/parallel/src/jitter.rs
@@ -1,53 +1,3 @@
-//! Debug jitter utilities for testing timing-related bugs.
-//!
-//! When the `debug-jitter` feature is enabled, proof workers can optionally add
-//! random delays after completing jobs. This helps trigger out-of-order timing bugs
-//! that may only manifest in real-world conditions.
-//!
-//! Control via environment variable:
-//! - `RETH_PROOF_WORKER_JITTER_MS`: Maximum jitter in milliseconds (0-N random delay)
-//!
-//! Example: `RETH_PROOF_WORKER_JITTER_MS=5` adds 0-5ms random delays.
+//! Re-export of debug jitter utilities from `reth-engine-primitives`.
 
-use std::{sync::OnceLock, thread, time::Duration};
-
-use rand::Rng;
-use tracing::trace;
-
-/// Cached jitter configuration from environment.
-static JITTER_CONFIG: OnceLock<Option<u64>> = OnceLock::new();
-
-/// Reads the jitter configuration from environment variables.
-///
-/// Returns `Some(max_ms)` if jitter is enabled, `None` otherwise.
-fn get_jitter_config() -> Option<u64> {
-    *JITTER_CONFIG.get_or_init(|| {
-        std::env::var("RETH_PROOF_WORKER_JITTER_MS")
-            .ok()
-            .and_then(|v| v.parse::<u64>().ok())
-            .filter(|&ms| ms > 0)
-    })
-}
-
-/// Applies a random jitter delay if configured.
-///
-/// When `RETH_PROOF_WORKER_JITTER_MS` is set to a positive value N,
-/// this function sleeps for a random duration between 0 and N milliseconds.
-///
-/// This is useful for testing timing-sensitive code paths that may have
-/// race conditions or ordering bugs that only manifest with variable latencies.
-pub(crate) fn maybe_apply_jitter(worker_id: usize, job_type: &str) {
-    if let Some(max_ms) = get_jitter_config() {
-        let delay_ms = rand::rng().random_range(0..=max_ms);
-        if delay_ms > 0 {
-            trace!(
-                target: "trie::proof_task",
-                worker_id,
-                job_type,
-                delay_ms,
-                "Applying debug jitter delay"
-            );
-            thread::sleep(Duration::from_millis(delay_ms));
-        }
-    }
-}
+pub(crate) use reth_engine_primitives::jitter::maybe_apply_jitter;
